@@ -1,9 +1,7 @@
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
-import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/catch';
-import 'rxjs/add/observable/throw';
+import { Observable } from 'rxjs';
+import { map, catchError } from 'rxjs/operators';
 
 @Injectable()
 export class NgxBraintreeService {
@@ -13,21 +11,19 @@ export class NgxBraintreeService {
   getClientToken(clientTokenURL: string): Observable<string> {
     return this.http
       .get(clientTokenURL, { responseType: 'json' })
-      .map((response: any) => {
-        return response.token;
-      })
-      .catch((error) => {
-        return Observable.throw(error);
-      });
+      .pipe(
+        map((response: any) => { return response.token; }),
+        catchError((error) => { return Observable.throw(error); })
+      );
   }
 
   createPurchase(createPurchaseURL: string, nonce: string, chargeAmount: number): Observable<any> {
     const headers = new HttpHeaders({ 'Content-Type': 'application/json' });
     return this.http
-    .post(createPurchaseURL, { nonce: nonce, chargeAmount: chargeAmount }, { 'headers': headers })
-      .map((response: any) => {
+      .post(createPurchaseURL, { nonce: nonce, chargeAmount: chargeAmount }, { 'headers': headers })
+      .pipe(map((response: any) => {
         return response;
-      });
+      }));
   }
 
 }
